refactor: color optim
