+++ react-materials/scaffolds/ice-customer-analysis-admin/README.md
@@ -2,12 +2,12 @@
 
 > 使用文档
 
-使用:
+## 使用:
 
 - 启动调试服务: `npm start`
 - 构建 dist: `npm run build`
 
-目录结构:
+## 目录结构:
 
 - react-router @4.x 默认采用 hashHistory 的单页应用
 - 入口文件: `src/index.js`
@@ -18,6 +18,12 @@
 - 通用组件: `src/components`
 - 页面文件: `src/pages`
 
-效果图:
+## 配色
+
+- 头部背景主色：#2f323f
+- 功能主色：#5e83fb
+- 字体颜色：#333、#666
+
+##效果图:
 
 ![screenshot](https://img.alicdn.com/tfs/TB1V.FZg3TqK1RjSZPhXXXfOFXa-2860-1520.png)
