+++ react-materials/scaffolds/ice-customer-analysis-admin/src/layouts/BasicLayout/BasicLayout.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
 import cx from 'classnames';
 import Layout from '@icedesign/layout';
-import Header from '../../components/Header';
-import Aside from '../../components/Aside';
-import Footer from '../../components/Footer';
+import Header from './components/Header';
+import Aside from './components/Aside';
+import Footer from './components/Footer';
 
 import './BasicLayout.scss';
