+++ react-materials/scaffolds/ice-customer-analysis-admin/src/layouts/BasicLayout/components/Header/Header.jsx
@@ -4,7 +4,7 @@ import { Balloon, Icon } from '@icedesign/base';
 import Menu, { SubMenu, Item as MenuItem } from '@icedesign/menu';
 import FoundationSymbol from 'foundation-symbol';
 import IceImg from '@icedesign/img';
-import { headerMenuConfig } from '../../menuConfig';
+import { headerMenuConfig } from '../../../../menuConfig';
 import Logo from '../Logo';
 import './Header.scss';
