+++ react-materials/scaffolds/ice-customer-analysis-admin/src/pages/Analysis/components/UserPortrait/AgeChart.jsx
@@ -44,9 +44,9 @@ class Columnmarked extends React.Component {
       },
     ];
     const colorSet = {
-      YOUNG: '#1890ff',
-      MIDDLE: '#f04864',
-      OLD: '#FED46B',
+      YOUNG: '#5e83fb',
+      MIDDLE: '#ee706d',
+      OLD: '#58ca9a',
     };
     Shape.registerShape('interval', 'textInterval', {
       drawShape(cfg, group) {
